trying to code wall jump
I 've been trying to implement the wall jump to my game. even tho it sounds simple for some reason I cant get it right.

I will work on another mechanics and leave the wall jumper to later. This update is a reminder that I will be trying to approach it next using  OnControllerColliderHit